move stake modal to folder
